Move around some types to more central locations in preparation.
