+++ src/types/types.ts
@@ -1,6 +1,6 @@
 import type { ComponentProps, ComponentType, SvelteComponent } from 'svelte';
 import type { Item5e, ItemCardContentComponent, ItemChatData } from './item';
-import type { RenderScheme } from 'src/api/api.types';
+import type { OnRenderArgs, RenderScheme } from 'src/api/api.types';
 
 export type Actor5e = any;
 
@@ -21,6 +21,29 @@ export type HtmlTabContent = {
   renderScheme: RenderScheme;
 };
 
+export type OnRenderTabArgs = OnRenderArgs & {
+  tabContentsElement: HTMLElement;
+};
+
+// TODO: Can this be removed?
+export type CustomTabBase = {
+  onRender?: (args: OnRenderTabArgs) => void;
+};
+
+// TODO: Can this be removed?
+export type CustomTab = CustomHtmlTab; // & others as time goes on
+
+// TODO: Can this be removed?
+export type CustomHtmlTab = {
+  type: 'html';
+  title: string;
+  tabId: string;
+  tabClasses: string[];
+  contentHtml: string;
+  tabContentsClasses: string[];
+  renderScheme: RenderScheme;
+} & CustomTabBase;
+
 // TODO: Make this generic in such a way that correct props are actually required and that an array of tabs can have hetergeneity of component types without a crazy TS type
 export type Tab<
   T extends SvelteComponent<any, any, any> = SvelteComponent<any, any, any>
@@ -28,6 +51,7 @@ export type Tab<
   id: string;
   displayName: string;
   content: SvelteTabContent<T> | HtmlTabContent;
+  onRender?: (args: OnRenderTabArgs) => void;
 };
 
 export type ClassSummary = {
